Update Package.swift to open/build under Xcode 11 (11.2.1)
When opening the package directly in Xcode 11, the IDE will give an error related to the tests being at a different path than standard. On building tests with the default target of macOS (since packages will support all platforms if not limited by supportedPlatforms) an error is given about the iso date format, since this was added in macOS 10.12.

This allows the package to open and build, although there are numerous test failures.